City forecast solution (#1)
* Created a proper stub for code development

Written up the relevant test scenarios in Gherkin.
Added entries in the README file.
Structured the project to make more sense in the next phase.

* Project file saved

* Added request handler and more models

The steps should now do something.

The appid secret is handled via dotenv.

* Fixed the request handler

Added .env template file. Needs instructions in the README for setup.

The test suite seems to treat SolutionDir/TestResults as its running
directory so that's where the .env file needs to land. Post-build event
added for that.

* Mapped the API response object

Refactored the request handler into something much simpler.

Extracted the environment handler to have a common place for additional
variables.

TODO: process the output of the API call and present the results in the
console

* Refactored the Gherkin scenario for 5 cities

Using the scenario outline to save on feature file space.

* Renamed before hook and refactored the paris scenario

* Code refactoring

* Added FluentAssertions and refactored code

* Logic stubs added

More code refactored

* Finalized data presentation

Need to calculate the values to feed into the presented string.

* Finalised the code logic

Displaying the correct values for the given locations in the console.

TODO: write documentation and create PR

* Adding screenshots and expanding the README

* Finalised docs and answered remaining questions

* Fixed README formatting

* Added missing setup step